+++ packages/_private/e2e-demo-test/src/test.selector.ts
@@ -38,6 +38,8 @@ const AllComponentSelectors = {
   combobox1036Delimiter: '#tab-content-Combobox syn-combobox[data-testid="combobox-1036-subsequently-changed-delimiter"]',
   combobox1056DelimiterAsyncPreValue: '#tab-content-Combobox syn-combobox[data-testid="combobox-1056-async-delimiter-change-with-async-pre-value"]',
   combobox1056DelimiterPreValue: '#tab-content-Combobox syn-combobox[data-testid="combobox-1056-async-delimiter-change-with-pre-value"]',
+  combobox1265DynamicOptionChangeButton: '#tab-content-Combobox syn-button[data-testid="combobox-1265-dynamic-option-button"]',
+  combobox1265DynamicOptionChanges: '#tab-content-Combobox syn-combobox[data-testid="combobox-1265-dynamic-option-changes"]',
   comboboxContent: '#tab-content-Combobox',
   comboboxFormReset: '#tab-content-Combobox form syn-button',
   comboboxLink: '#tab-Combobox',
@@ -76,6 +78,10 @@ const AllComponentSelectors = {
   optgroupThirdEnabledItems: '#tab-content-OptGroup syn-select syn-optgroup:nth-of-type(3) syn-option:not([disabled])',
   optgroupThirdItem: '#tab-content-OptGroup syn-select syn-optgroup:nth-of-type(3)',
 
+  // Range
+  range1272ChangeValueButton: '#tab-content-Range syn-button[data-testid="range-1272-change-value-button"]',
+  range1272ProgramaticValueChange: '#tab-content-Range syn-range[data-testid="range-1272-programatic-value-change"]',
+
   // Select
   select847Multiple: '#tab-content-Select syn-select[data-testid="select-847-multiple"]',
   select885ValueZeroNumber: '#tab-content-Select syn-select[data-testid="select-885-value-zero-number"]',
@@ -83,6 +89,8 @@ const AllComponentSelectors = {
   select1036Delimiter: '#tab-content-Select syn-select[data-testid="select-1036-subsequently-changed-delimiter"]',
   select1056DelimiterAsyncPreValue: '#tab-content-Select syn-select[data-testid="select-1056-async-delimiter-change-with-async-pre-value"]',
   select1056DelimiterPreValue: '#tab-content-Select syn-select[data-testid="select-1056-async-delimiter-change-with-pre-value"]',
+  select1265DynamicOptionChangeButton: '#tab-content-Select syn-button[data-testid="select-1265-dynamic-option-button"]',
+  select1265DynamicOptionChanges: '#tab-content-Select syn-select[data-testid="select-1265-dynamic-option-changes"]',
   selectContent: '#tab-content-Select',
   selectForm: '#tab-content-Select syn-select[data-testid="select-form-813"]',
   selectFormOptions: '#tab-content-Select syn-select[data-testid="select-form-813"] syn-option',