+++ packages/_private/vanilla-demo/src/AllComponentParts/Combobox.ts
@@ -1,4 +1,5 @@
 import { html } from 'lit';
+import { updateComboboxRegressions1265 } from '@synergy-design-system/demo-utilities';
 import type { RegressionFns } from '../all-components-regressions';
 
 export const Combobox = (regressions: RegressionFns = []) => {
@@ -128,5 +129,23 @@ export const Combobox = (regressions: RegressionFns = []) => {
       <syn-option value="option-2">Option 2</syn-option>
       <syn-option value="option-3">Option 3</syn-option>
     </syn-combobox>
+
+    <syn-combobox
+      data-testid="combobox-1265-dynamic-option-changes"
+      label="Dynamic Option Changes"
+      restricted
+      value="option-2"
+    >
+      <syn-option value="option-1">Option 1</syn-option>
+      <syn-option value="option-2">Option 2</syn-option>
+      <syn-option value="option-3">Option 2</syn-option>
+    </syn-combobox>
+
+    <syn-button
+      data-testid="combobox-1265-dynamic-option-button"
+      @click=${updateComboboxRegressions1265}
+    >
+      Dynamically change option 2 to "Changed Option 2"
+    </syn-button>
   `;
 };